Perform nightly releases if tests pass with the latest supported Django.
We used to only perform a nightly release if our tests passed with Django SVN.
This was more important back in the Django <= 1.0 days, but less so now.
Now nightlies are built and released so long as the latest supported/tested
release of Django works. We don't guarantee things will work with Django SVN.